requirements.txt
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-core>=0.1.15
langchain-community>=0.0.28
langgraph>=0.0.20
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
flask>=2.0.0
gunicorn>=20.1.0
pydantic>=2.0.0
typing-extensions>=4.5.0
redis>=4.5.1
structlog>=23.1.0
numpy>=1.24.0
prometheus-client>=0.16.0
